+++ src/test/java/com/sprint/deokhugamteam7/domain/notification/unit/entity/NotificationTest.java
@@ -1,44 +1,24 @@
-package com.sprint.deokhugamteam7.domain.notification.service.impl;
-
+package com.sprint.deokhugamteam7.domain.notification.unit.entity;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import com.sprint.deokhugamteam7.constant.NotificationType;
 import com.sprint.deokhugamteam7.domain.book.entity.Book;
-import com.sprint.deokhugamteam7.domain.notification.dto.NotificationDto;
-import com.sprint.deokhugamteam7.domain.notification.dto.NotificationUpdateRequest;
+import com.sprint.deokhugamteam7.domain.comment.entity.Comment;
 import com.sprint.deokhugamteam7.domain.notification.entity.Notification;
-import com.sprint.deokhugamteam7.domain.notification.repository.NotificationRepository;
-import com.sprint.deokhugamteam7.domain.notification.service.NotificationServiceImpl;
 import com.sprint.deokhugamteam7.domain.review.entity.Review;
 import com.sprint.deokhugamteam7.domain.user.entity.User;
-import com.sprint.deokhugamteam7.domain.user.repository.UserRepository;
 import com.sprint.deokhugamteam7.exception.ErrorCode;
 import com.sprint.deokhugamteam7.exception.notification.NotificationException;
 import java.time.LocalDate;
-import java.util.Optional;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 
-@ExtendWith(MockitoExtension.class)
-class NotificationServiceImplTest {
-
-    @InjectMocks
-    private NotificationServiceImpl notificationService;
-
-    @Mock
-    private NotificationRepository notificationRepository;
-
-    @Mock
-    private UserRepository userRepository;
+class NotificationTest {
 
     private final UUID NOTIFICATION_ID = UUID.fromString("11111111-1111-1111-1111-111111111111");
     private final UUID USER_ID = UUID.fromString("22222222-2222-2222-2222-222222222222");
@@ -48,7 +28,6 @@ class NotificationServiceImplTest {
     private User otherUser;
     private Review review;
     private Notification notification;
-    private NotificationUpdateRequest request;
 
     @BeforeEach
     void setup() {
@@ -65,36 +44,41 @@ class NotificationServiceImplTest {
 
         notification = Notification.create(user, review, "테스트 알림");
         setPrivateField(notification, "id", NOTIFICATION_ID);
-
-        request = new NotificationUpdateRequest(true);
     }
 
     @Test
-    @DisplayName("알림 수정 성공 - confirmed 수정 확인")
-    void 알림_수정_성공() {
-        // given
-        given(notificationRepository.findById(any())).willReturn(Optional.of(notification));
-        given(userRepository.findById(any())).willReturn(Optional.of(user));
+    @DisplayName("알림 생성 성공 - Like")
+    void createNotificationWithNotificationTypeToLike() {
+        Notification likeNotification = Notification.create(user, review, NotificationType.LIKE.formatMessage(otherUser, null));
 
-        // when
-        NotificationDto result = notificationService.update(NOTIFICATION_ID, USER_ID, request);
-
-        // then
-        assertThat(result.confirmed()).isTrue();
+        assertThat(likeNotification.getContent()).isEqualTo("[test2]님이 나의 리뷰를 좋아합니다.");
+        assertThat(likeNotification.isDelete()).isFalse();
     }
 
     @Test
-    @DisplayName(("알림 수정 실패 - NOTIFICATION_NOT_FOUND"))
-    void notificationNotFoundById() {
-        // given
-        given(notificationRepository.findById(notification.getId())).willReturn(Optional.empty());
+    @DisplayName("알림 생성 성공 - Like")
+    void createNotificationWithNotificationTypeToComment() {
+        Comment comment = Comment.create(otherUser, review, "댓글 테스트"); // 에러 발생 부분
 
-        // when & then
-        assertThatThrownBy(() -> notificationService.update(notification.getId(), user.getId(), request))
+        Notification likeNotification = Notification.create(user, review, NotificationType.COMMENT.formatMessage(otherUser, comment));
+
+        assertThat(likeNotification.getContent()).isEqualTo("[test2]님이 나의 리뷰에 댓글을 남겼습니다.\n댓글 테스트");
+        assertThat(likeNotification.isDelete()).isFalse();
+    }
+
+    @Test
+    @DisplayName("알림 접근 권한 검사 성공 - 작성자와 요청자 일치")
+    void validateUserAuthorization_success() {
+        assertDoesNotThrow(() -> notification.validateUserAuthorization(USER_ID));
+    }
+
+    @Test
+    @DisplayName("알림 접근 권한 검사 실패 - 작성자와 요청자 불일치")
+    void validateUserAuthorization_fail() {
+        assertThatThrownBy(() -> notification.validateUserAuthorization(OTHER_USER_ID))
             .isInstanceOf(NotificationException.class)
-            .hasMessageContaining(ErrorCode.NOTIFICATION_NOT_FOUND.getMessage());
+            .hasMessageContaining(ErrorCode.NOTIFICATION_NOT_OWNED.getMessage());
     }
-
 
     private void setPrivateField(Object target, String fieldName, Object value) {
         try {
@@ -105,4 +89,5 @@ class NotificationServiceImplTest {
             throw new RuntimeException("리플렉션으로 필드 설정 실패: " + fieldName, e);
         }
     }
+
 }
